pyproject.toml
[project]
name = "rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "google-genai>=1.31.0",
    "groq>=0.31.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]
